New SETTING: TREAT_TIMEOUTS_AS_DISCONNECT -Set to 1 (true) by default -If Internet test gets (number:fluke_test_eliminator) timeouts in a row, will treat as either connected or disconnected, depending on what this setting is set to.
The routers at my school will semi-kick-off devices. In this state this program gets timeouts when it tries to connect to different websites. The program fixes this with one of it's 4 fixes. 

However, some places require a different kind of browser network authentication. While not logged in, the program gets timeouts when it tests the connection. Users who have this type of network authentication will need to set this setting to 0 (False)