Internet Test was happening too fast in some cases and not allowing enough time for the network adapter to connect. These Added check to make sure the Internet Connection test lasts at least %T_MILI_SMALLEST% (200) milliseconds if the test result is 0 (not connected).
Internet Test was happening too fast in some
cases and not allowing enough time for the 
network adapter to connect. These tests would
last less than 2 seconds (at least on my machine)
and result in 5 "Could not connect" messages
in a row, thus starting or restarting the 
process to fix the connection. 

If the script detects that the test was too fast,
it will stall for 3 seconds and try again. It will
stall for a maximum of %MaxStall% (10) times before 
treating the connection as one that needs to be fixed.