Feature: Re-added Automatic selection of Network Connection name if OMIT_USER_INPUT is enabled. It makes it's best guess as to what the network connection should be. This really shouldn't be used though. It might work in some cases, but definitely not all.
It checks first to see if "Wireless Network Connection [10-1]" exist. If one does, it selects the first one it finds (It goes from 10 to nothing, So it checks "Wireless Network Connection 10" then "Wireless Network Connection 9"etc. 

If none of those exist, It then does the same thing except with "Local Area Connection"

If none of those exist, it selects the first one that is output from "NETSH INTERFACE SHOW INTERFACE"

This is by no means guaranteed to select the right Network Connection. It is merely a best guess to allow the program to continue to function if the current value for NETWORK is not valid and there is no user input.