Re-organize Settings
This should the settings easier to find and the settings menus easier to navigate.

There are now 5 setting categories:
-Main Settings
-Display Settings
-Update Settings
-Fixes Configuration (enable/disable fixes and change fix-specific settings)
-Advanced Settings


Some code cleanup
Change the way settings are loaded and applied
-No longer loads defaults and then changes the custom settings
-Intead it sets NUL values to the settings, loads custom settings, and checks to make sure they all have a value. 
This way, new settings that are not yet set can have their default values dependent on current user-defined settings.

For example: There's a new fix that affects all network connections, similar to the USE_IP_RESET fix. Since anyone who disables the USE_IP_RESET fix would probably not want this new fix enabled, the new fixes' default values are dependent on the value of the USE_IP_RESET fix. On the other hand, if someone wants the global fixes to be used, the default value for the new fix will be enabled. 

In this way the automatic update will not cause significant change in the behavior of the program as customized by the user.